Text composer: start extracting in a dedicated ViewModel/State/Action/Events
