+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailAction.kt
@@ -30,10 +30,7 @@ import org.matrix.android.sdk.api.session.widgets.model.Widget
 import org.matrix.android.sdk.api.util.MatrixItem
 
 sealed class RoomDetailAction : VectorViewModelAction {
-    data class UserIsTyping(val isTyping: Boolean) : RoomDetailAction()
-    data class SaveDraft(val draft: String) : RoomDetailAction()
     data class SendSticker(val stickerContent: MessageStickerContent) : RoomDetailAction()
-    data class SendMessage(val text: CharSequence, val autoMarkdown: Boolean) : RoomDetailAction()
     data class SendMedia(val attachments: List<ContentAttachmentData>, val compressBeforeSending: Boolean) : RoomDetailAction()
     data class TimelineEventTurnsVisible(val event: TimelineEvent) : RoomDetailAction()
     data class TimelineEventTurnsInvisible(val event: TimelineEvent) : RoomDetailAction()
@@ -51,11 +48,6 @@ sealed class RoomDetailAction : VectorViewModelAction {
 
     object EnterTrackingUnreadMessagesState : RoomDetailAction()
     object ExitTrackingUnreadMessagesState : RoomDetailAction()
-
-    data class EnterEditMode(val eventId: String, val text: String) : RoomDetailAction()
-    data class EnterQuoteMode(val eventId: String, val text: String) : RoomDetailAction()
-    data class EnterReplyMode(val eventId: String, val text: String) : RoomDetailAction()
-    data class EnterRegularMode(val text: String, val fromSharing: Boolean) : RoomDetailAction()
 
     data class ResendMessage(val eventId: String) : RoomDetailAction()
     data class RemoveFailedEcho(val eventId: String) : RoomDetailAction()
