+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailViewModel.kt
@@ -44,8 +44,6 @@ import im.vector.app.features.call.conference.JitsiActiveConferenceHolder
 import im.vector.app.features.call.conference.JitsiService
 import im.vector.app.features.call.lookup.CallProtocolsChecker
 import im.vector.app.features.call.webrtc.WebRtcCallManager
-import im.vector.app.features.command.CommandParser
-import im.vector.app.features.command.ParsedCommand
 import im.vector.app.features.createdirect.DirectRoomHelper
 import im.vector.app.features.crypto.keysrequest.OutboundSessionKeySharingStrategy
 import im.vector.app.features.crypto.verification.SupportedVerificationMethodsProvider
@@ -67,9 +65,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import org.commonmark.parser.Parser
-import org.commonmark.renderer.html.HtmlRenderer
-import org.matrix.android.sdk.api.MatrixCallback
 import org.matrix.android.sdk.api.MatrixPatterns
 import org.matrix.android.sdk.api.extensions.tryOrNull
 import org.matrix.android.sdk.api.query.QueryStringValue
@@ -86,22 +81,14 @@ import org.matrix.android.sdk.api.session.initsync.SyncStatusService
 import org.matrix.android.sdk.api.session.room.members.ChangeMembershipState
 import org.matrix.android.sdk.api.session.room.members.roomMemberQueryParams
 import org.matrix.android.sdk.api.session.room.model.Membership
-import org.matrix.android.sdk.api.session.room.model.PowerLevelsContent
 import org.matrix.android.sdk.api.session.room.model.RoomMemberSummary
 import org.matrix.android.sdk.api.session.room.model.RoomSummary
-import org.matrix.android.sdk.api.session.room.model.message.MessageType
-import org.matrix.android.sdk.api.session.room.model.message.OptionItem
 import org.matrix.android.sdk.api.session.room.model.message.getFileUrl
 import org.matrix.android.sdk.api.session.room.model.tombstone.RoomTombstoneContent
 import org.matrix.android.sdk.api.session.room.powerlevels.PowerLevelsHelper
 import org.matrix.android.sdk.api.session.room.read.ReadService
-import org.matrix.android.sdk.api.session.room.send.UserDraft
 import org.matrix.android.sdk.api.session.room.timeline.Timeline
 import org.matrix.android.sdk.api.session.room.timeline.TimelineEvent
-import org.matrix.android.sdk.api.session.room.timeline.getLastMessageContent
-import org.matrix.android.sdk.api.session.room.timeline.getRelationContent
-import org.matrix.android.sdk.api.session.room.timeline.getTextEditableContent
-import org.matrix.android.sdk.api.session.space.CreateSpaceParams
 import org.matrix.android.sdk.api.session.widgets.model.WidgetType
 import org.matrix.android.sdk.api.util.toOptional
 import org.matrix.android.sdk.internal.crypto.model.event.WithHeldCode
@@ -181,7 +168,6 @@ class RoomDetailViewModel @AssistedInject constructor(
         observeSyncState()
         observeDataStore()
         observeEventDisplayedActions()
-        loadDraftIfAny()
         observeUnreadState()
         observeMyRoomMember()
         observeActiveRoomWidgets()
@@ -235,13 +221,11 @@ class RoomDetailViewModel @AssistedInject constructor(
     private fun observePowerLevel() {
         PowerLevelsObservableFactory(room).createObservable()
                 .subscribe {
-                    val canSendMessage = PowerLevelsHelper(it).isUserAllowedToSend(session.myUserId, false, EventType.MESSAGE)
                     val canInvite = PowerLevelsHelper(it).isUserAbleToInvite(session.myUserId)
                     val isAllowedToManageWidgets = session.widgetService().hasPermissionsToHandleWidgets(room.roomId)
                     val isAllowedToStartWebRTCCall = PowerLevelsHelper(it).isUserAllowedToSend(session.myUserId, false, EventType.CALL_INVITE)
                     setState {
                         copy(
-                                canSendMessage = canSendMessage,
                                 canInvite = canInvite,
                                 isAllowedToManageWidgets = isAllowedToManageWidgets,
                                 isAllowedToStartWebRTCCall = isAllowedToStartWebRTCCall
@@ -300,10 +284,7 @@ class RoomDetailViewModel @AssistedInject constructor(
 
     override fun handle(action: RoomDetailAction) {
         when (action) {
-            is RoomDetailAction.UserIsTyping                     -> handleUserIsTyping(action)
             is RoomDetailAction.ComposerFocusChange              -> handleComposerFocusChange(action)
-            is RoomDetailAction.SaveDraft                        -> handleSaveDraft(action)
-            is RoomDetailAction.SendMessage                      -> handleSendMessage(action)
             is RoomDetailAction.SendMedia                        -> handleSendMedia(action)
             is RoomDetailAction.SendSticker                      -> handleSendSticker(action)
             is RoomDetailAction.TimelineEventTurnsVisible        -> handleEventVisible(action)
@@ -315,10 +296,6 @@ class RoomDetailViewModel @AssistedInject constructor(
             is RoomDetailAction.RedactAction                     -> handleRedactEvent(action)
             is RoomDetailAction.UndoReaction                     -> handleUndoReact(action)
             is RoomDetailAction.UpdateQuickReactAction           -> handleUpdateQuickReaction(action)
-            is RoomDetailAction.EnterRegularMode                 -> handleEnterRegularMode(action)
-            is RoomDetailAction.EnterEditMode                    -> handleEditAction(action)
-            is RoomDetailAction.EnterQuoteMode                   -> handleQuoteAction(action)
-            is RoomDetailAction.EnterReplyMode                   -> handleReplyAction(action)
             is RoomDetailAction.DownloadOrOpen                   -> handleOpenOrDownloadFile(action)
             is RoomDetailAction.NavigateToEvent                  -> handleNavigateToEvent(action)
             is RoomDetailAction.JoinAndOpenReplacementRoom       -> handleJoinAndOpenReplacementRoom()
@@ -590,70 +567,6 @@ class RoomDetailViewModel @AssistedInject constructor(
         return room.getRoomMember(userId)
     }
 
-    /**
-     * Convert a send mode to a draft and save the draft
-     */
-    private fun handleSaveDraft(action: RoomDetailAction.SaveDraft) = withState {
-        session.coroutineScope.launch {
-            when {
-                it.sendMode is SendMode.REGULAR && !it.sendMode.fromSharing -> {
-                    setState { copy(sendMode = it.sendMode.copy(action.draft)) }
-                    room.saveDraft(UserDraft.REGULAR(action.draft))
-                }
-                it.sendMode is SendMode.REPLY                               -> {
-                    setState { copy(sendMode = it.sendMode.copy(text = action.draft)) }
-                    room.saveDraft(UserDraft.REPLY(it.sendMode.timelineEvent.root.eventId!!, action.draft))
-                }
-                it.sendMode is SendMode.QUOTE                               -> {
-                    setState { copy(sendMode = it.sendMode.copy(text = action.draft)) }
-                    room.saveDraft(UserDraft.QUOTE(it.sendMode.timelineEvent.root.eventId!!, action.draft))
-                }
-                it.sendMode is SendMode.EDIT                                -> {
-                    setState { copy(sendMode = it.sendMode.copy(text = action.draft)) }
-                    room.saveDraft(UserDraft.EDIT(it.sendMode.timelineEvent.root.eventId!!, action.draft))
-                }
-            }
-        }
-    }
-
-    private fun loadDraftIfAny() {
-        val currentDraft = room.getDraft()
-        setState {
-            copy(
-                    // Create a sendMode from a draft and retrieve the TimelineEvent
-                    sendMode = when (currentDraft) {
-                        is UserDraft.REGULAR -> SendMode.REGULAR(currentDraft.text, false)
-                        is UserDraft.QUOTE   -> {
-                            room.getTimeLineEvent(currentDraft.linkedEventId)?.let { timelineEvent ->
-                                SendMode.QUOTE(timelineEvent, currentDraft.text)
-                            }
-                        }
-                        is UserDraft.REPLY   -> {
-                            room.getTimeLineEvent(currentDraft.linkedEventId)?.let { timelineEvent ->
-                                SendMode.REPLY(timelineEvent, currentDraft.text)
-                            }
-                        }
-                        is UserDraft.EDIT    -> {
-                            room.getTimeLineEvent(currentDraft.linkedEventId)?.let { timelineEvent ->
-                                SendMode.EDIT(timelineEvent, currentDraft.text)
-                            }
-                        }
-                        else                 -> null
-                    } ?: SendMode.REGULAR("", fromSharing = false)
-            )
-        }
-    }
-
-    private fun handleUserIsTyping(action: RoomDetailAction.UserIsTyping) {
-        if (vectorPreferences.sendTypingNotifs()) {
-            if (action.isTyping) {
-                room.userIsTyping()
-            } else {
-                room.userStopsTyping()
-            }
-        }
-    }
-
     private fun handleComposerFocusChange(action: RoomDetailAction.ComposerFocusChange) {
         // Ensure outbound session keys
         if (OutboundSessionKeySharingStrategy.WhenTyping == BuildConfig.outboundSessionKeySharingStrategy && room.isEncrypted()) {
@@ -766,417 +679,7 @@ class RoomDetailViewModel @AssistedInject constructor(
         }
     }
 
-// PRIVATE METHODS *****************************************************************************
-
-    private fun handleSendMessage(action: RoomDetailAction.SendMessage) {
-        withState { state ->
-            when (state.sendMode) {
-                is SendMode.REGULAR -> {
-                    when (val slashCommandResult = CommandParser.parseSplashCommand(action.text)) {
-                        is ParsedCommand.ErrorNotACommand         -> {
-                            // Send the text message to the room
-                            room.sendTextMessage(action.text, autoMarkdown = action.autoMarkdown)
-                            _viewEvents.post(RoomDetailViewEvents.MessageSent)
-                            popDraft()
-                        }
-                        is ParsedCommand.ErrorSyntax              -> {
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandError(slashCommandResult.command))
-                        }
-                        is ParsedCommand.ErrorEmptySlashCommand   -> {
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandUnknown("/"))
-                        }
-                        is ParsedCommand.ErrorUnknownSlashCommand -> {
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandUnknown(slashCommandResult.slashCommand))
-                        }
-                        is ParsedCommand.SendPlainText            -> {
-                            // Send the text message to the room, without markdown
-                            room.sendTextMessage(slashCommandResult.message, autoMarkdown = false)
-                            _viewEvents.post(RoomDetailViewEvents.MessageSent)
-                            popDraft()
-                        }
-                        is ParsedCommand.Invite                   -> {
-                            handleInviteSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.Invite3Pid               -> {
-                            handleInvite3pidSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.SetUserPowerLevel        -> {
-                            handleSetUserPowerLevel(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.ClearScalarToken         -> {
-                            // TODO
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandNotImplemented)
-                        }
-                        is ParsedCommand.SetMarkdown              -> {
-                            vectorPreferences.setMarkdownEnabled(slashCommandResult.enable)
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled(
-                                    if (slashCommandResult.enable) R.string.markdown_has_been_enabled else R.string.markdown_has_been_disabled))
-                            popDraft()
-                        }
-                        is ParsedCommand.UnbanUser                -> {
-                            handleUnbanSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.BanUser                  -> {
-                            handleBanSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.KickUser                 -> {
-                            handleKickSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.JoinRoom                 -> {
-                            handleJoinToAnotherRoomSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.PartRoom                 -> {
-                            // TODO
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandNotImplemented)
-                        }
-                        is ParsedCommand.SendEmote                -> {
-                            room.sendTextMessage(slashCommandResult.message, msgType = MessageType.MSGTYPE_EMOTE, autoMarkdown = action.autoMarkdown)
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendRainbow              -> {
-                            slashCommandResult.message.toString().let {
-                                room.sendFormattedTextMessage(it, rainbowGenerator.generate(it))
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendRainbowEmote         -> {
-                            slashCommandResult.message.toString().let {
-                                room.sendFormattedTextMessage(it, rainbowGenerator.generate(it), MessageType.MSGTYPE_EMOTE)
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendSpoiler              -> {
-                            room.sendFormattedTextMessage(
-                                    "[${stringProvider.getString(R.string.spoiler)}](${slashCommandResult.message})",
-                                    "<span data-mx-spoiler>${slashCommandResult.message}</span>"
-                            )
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendShrug                -> {
-                            val sequence = buildString {
-                                append("¯\\_(ツ)_/¯")
-                                if (slashCommandResult.message.isNotEmpty()) {
-                                    append(" ")
-                                    append(slashCommandResult.message)
-                                }
-                            }
-                            room.sendTextMessage(sequence)
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendChatEffect           -> {
-                            sendChatEffect(slashCommandResult)
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.SendPoll                 -> {
-                            room.sendPoll(slashCommandResult.question, slashCommandResult.options.mapIndexed { index, s -> OptionItem(s, "$index. $s") })
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.ChangeTopic              -> {
-                            handleChangeTopicSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.ChangeDisplayName        -> {
-                            handleChangeDisplayNameSlashCommand(slashCommandResult)
-                            popDraft()
-                        }
-                        is ParsedCommand.DiscardSession           -> {
-                            if (room.isEncrypted()) {
-                                session.cryptoService().discardOutboundSession(room.roomId)
-                                _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                                popDraft()
-                            } else {
-                                _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                                _viewEvents.post(
-                                        RoomDetailViewEvents
-                                                .ShowMessage(stringProvider.getString(R.string.command_description_discard_session_not_handled))
-                                )
-                            }
-                        }
-                        is ParsedCommand.CreateSpace              -> {
-                            viewModelScope.launch(Dispatchers.IO) {
-                                try {
-                                    val params = CreateSpaceParams().apply {
-                                        name = slashCommandResult.name
-                                        invitedUserIds.addAll(slashCommandResult.invitees)
-                                    }
-                                    val spaceId = session.spaceService().createSpace(params)
-                                    session.spaceService().getSpace(spaceId)
-                                            ?.addChildren(
-                                                    state.roomId,
-                                                    null,
-                                                    null,
-                                                    true
-                                            )
-                                } catch (failure: Throwable) {
-                                    _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-                                }
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.AddToSpace               -> {
-                            viewModelScope.launch(Dispatchers.IO) {
-                                try {
-                                    session.spaceService().getSpace(slashCommandResult.spaceId)
-                                            ?.addChildren(
-                                                    room.roomId,
-                                                    null,
-                                                    null,
-                                                    false
-                                            )
-                                } catch (failure: Throwable) {
-                                    _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-                                }
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.JoinSpace                -> {
-                            viewModelScope.launch(Dispatchers.IO) {
-                                try {
-                                    session.spaceService().joinSpace(slashCommandResult.spaceIdOrAlias)
-                                } catch (failure: Throwable) {
-                                    _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-                                }
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.LeaveRoom                -> {
-                            viewModelScope.launch(Dispatchers.IO) {
-                                try {
-                                    session.getRoom(slashCommandResult.roomId)?.leave(null)
-                                } catch (failure: Throwable) {
-                                    _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-                                }
-                            }
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                        is ParsedCommand.UpgradeRoom              -> {
-                            _viewEvents.post(
-                                    RoomDetailViewEvents.ShowRoomUpgradeDialog(
-                                            slashCommandResult.newVersion,
-                                            room.roomSummary()?.isPublic ?: false
-                                    )
-                            )
-                            _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-                            popDraft()
-                        }
-                    }.exhaustive
-                }
-                is SendMode.EDIT    -> {
-                    // is original event a reply?
-                    val inReplyTo = state.sendMode.timelineEvent.getRelationContent()?.inReplyTo?.eventId
-                    if (inReplyTo != null) {
-                        // TODO check if same content?
-                        room.getTimeLineEvent(inReplyTo)?.let {
-                            room.editReply(state.sendMode.timelineEvent, it, action.text.toString())
-                        }
-                    } else {
-                        val messageContent = state.sendMode.timelineEvent.getLastMessageContent()
-                        val existingBody = messageContent?.body ?: ""
-                        if (existingBody != action.text) {
-                            room.editTextMessage(state.sendMode.timelineEvent,
-                                    messageContent?.msgType ?: MessageType.MSGTYPE_TEXT,
-                                    action.text,
-                                    action.autoMarkdown)
-                        } else {
-                            Timber.w("Same message content, do not send edition")
-                        }
-                    }
-                    _viewEvents.post(RoomDetailViewEvents.MessageSent)
-                    popDraft()
-                }
-                is SendMode.QUOTE   -> {
-                    val messageContent = state.sendMode.timelineEvent.getLastMessageContent()
-                    val textMsg = messageContent?.body
-
-                    val finalText = legacyRiotQuoteText(textMsg, action.text.toString())
-
-                    // TODO check for pills?
-
-                    // TODO Refactor this, just temporary for quotes
-                    val parser = Parser.builder().build()
-                    val document = parser.parse(finalText)
-                    val renderer = HtmlRenderer.builder().build()
-                    val htmlText = renderer.render(document)
-                    if (finalText == htmlText) {
-                        room.sendTextMessage(finalText)
-                    } else {
-                        room.sendFormattedTextMessage(finalText, htmlText)
-                    }
-                    _viewEvents.post(RoomDetailViewEvents.MessageSent)
-                    popDraft()
-                }
-                is SendMode.REPLY   -> {
-                    state.sendMode.timelineEvent.let {
-                        room.replyToMessage(it, action.text.toString(), action.autoMarkdown)
-                        _viewEvents.post(RoomDetailViewEvents.MessageSent)
-                        popDraft()
-                    }
-                }
-            }.exhaustive
-        }
-    }
-
-    private fun sendChatEffect(sendChatEffect: ParsedCommand.SendChatEffect) {
-        // If message is blank, convert to an emote, with default message
-        if (sendChatEffect.message.isBlank()) {
-            val defaultMessage = stringProvider.getString(when (sendChatEffect.chatEffect) {
-                ChatEffect.CONFETTI -> R.string.default_message_emote_confetti
-                ChatEffect.SNOWFALL -> R.string.default_message_emote_snow
-            })
-            room.sendTextMessage(defaultMessage, MessageType.MSGTYPE_EMOTE)
-        } else {
-            room.sendTextMessage(sendChatEffect.message, sendChatEffect.chatEffect.toMessageType())
-        }
-    }
-
-    private fun popDraft() = withState {
-        if (it.sendMode is SendMode.REGULAR && it.sendMode.fromSharing) {
-            // If we were sharing, we want to get back our last value from draft
-            loadDraftIfAny()
-        } else {
-            // Otherwise we clear the composer and remove the draft from db
-            setState { copy(sendMode = SendMode.REGULAR("", false)) }
-            viewModelScope.launch {
-                room.deleteDraft()
-            }
-        }
-    }
-
-    private fun handleJoinToAnotherRoomSlashCommand(command: ParsedCommand.JoinRoom) {
-        viewModelScope.launch {
-            try {
-                session.joinRoom(command.roomAlias, command.reason, emptyList())
-            } catch (failure: Throwable) {
-                _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-                return@launch
-            }
-            session.getRoomSummary(command.roomAlias)
-                    ?.roomId
-                    ?.let {
-                        _viewEvents.post(RoomDetailViewEvents.JoinRoomCommandSuccess(it))
-                    }
-        }
-    }
-
-    private fun legacyRiotQuoteText(quotedText: String?, myText: String): String {
-        val messageParagraphs = quotedText?.split("\n\n".toRegex())?.dropLastWhile { it.isEmpty() }?.toTypedArray()
-        return buildString {
-            if (messageParagraphs != null) {
-                for (i in messageParagraphs.indices) {
-                    if (messageParagraphs[i].isNotBlank()) {
-                        append("> ")
-                        append(messageParagraphs[i])
-                    }
-
-                    if (i != messageParagraphs.lastIndex) {
-                        append("\n\n")
-                    }
-                }
-            }
-            append("\n\n")
-            append(myText)
-        }
-    }
-
-    private fun handleChangeTopicSlashCommand(changeTopic: ParsedCommand.ChangeTopic) {
-        launchSlashCommandFlowSuspendable {
-            room.updateTopic(changeTopic.topic)
-        }
-    }
-
-    private fun handleInviteSlashCommand(invite: ParsedCommand.Invite) {
-        launchSlashCommandFlowSuspendable {
-            room.invite(invite.userId, invite.reason)
-        }
-    }
-
-    private fun handleInvite3pidSlashCommand(invite: ParsedCommand.Invite3Pid) {
-        launchSlashCommandFlowSuspendable {
-            room.invite3pid(invite.threePid)
-        }
-    }
-
-    private fun handleSetUserPowerLevel(setUserPowerLevel: ParsedCommand.SetUserPowerLevel) {
-        val newPowerLevelsContent = room.getStateEvent(EventType.STATE_ROOM_POWER_LEVELS)
-                ?.content
-                ?.toModel<PowerLevelsContent>()
-                ?.setUserPowerLevel(setUserPowerLevel.userId, setUserPowerLevel.powerLevel)
-                ?.toContent()
-                ?: return
-
-        launchSlashCommandFlowSuspendable {
-            room.sendStateEvent(EventType.STATE_ROOM_POWER_LEVELS, null, newPowerLevelsContent)
-        }
-    }
-
-    private fun handleChangeDisplayNameSlashCommand(changeDisplayName: ParsedCommand.ChangeDisplayName) {
-        launchSlashCommandFlowSuspendable {
-            session.setDisplayName(session.myUserId, changeDisplayName.displayName)
-        }
-    }
-
-    private fun handleKickSlashCommand(kick: ParsedCommand.KickUser) {
-        launchSlashCommandFlowSuspendable {
-            room.kick(kick.userId, kick.reason)
-        }
-    }
-
-    private fun handleBanSlashCommand(ban: ParsedCommand.BanUser) {
-        launchSlashCommandFlowSuspendable {
-            room.ban(ban.userId, ban.reason)
-        }
-    }
-
-    private fun handleUnbanSlashCommand(unban: ParsedCommand.UnbanUser) {
-        launchSlashCommandFlowSuspendable {
-            room.unban(unban.userId, unban.reason)
-        }
-    }
-
-    private fun launchSlashCommandFlow(lambda: (MatrixCallback<Unit>) -> Unit) {
-        _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-        val matrixCallback = object : MatrixCallback<Unit> {
-            override fun onSuccess(data: Unit) {
-                _viewEvents.post(RoomDetailViewEvents.SlashCommandResultOk)
-            }
-
-            override fun onFailure(failure: Throwable) {
-                _viewEvents.post(RoomDetailViewEvents.SlashCommandResultError(failure))
-            }
-        }
-        lambda.invoke(matrixCallback)
-    }
-
-    private fun launchSlashCommandFlowSuspendable(block: suspend () -> Unit) {
-        _viewEvents.post(RoomDetailViewEvents.SlashCommandHandled())
-        viewModelScope.launch {
-            val event = try {
-                block()
-                RoomDetailViewEvents.SlashCommandResultOk
-            } catch (failure: Exception) {
-                RoomDetailViewEvents.SlashCommandResultError(failure)
-            }
-            _viewEvents.post(event)
-        }
-    }
+    // PRIVATE METHODS *****************************************************************************
 
     private fun handleSendReaction(action: RoomDetailAction.SendReaction) {
         room.sendReaction(action.targetEventId, action.reaction)
@@ -1244,28 +747,6 @@ class RoomDetailViewModel @AssistedInject constructor(
         viewModelScope.launch {
             tryOrNull { room.join() }
         }
-    }
-
-    private fun handleEditAction(action: RoomDetailAction.EnterEditMode) {
-        room.getTimeLineEvent(action.eventId)?.let { timelineEvent ->
-            setState { copy(sendMode = SendMode.EDIT(timelineEvent, timelineEvent.getTextEditableContent() ?: "")) }
-        }
-    }
-
-    private fun handleQuoteAction(action: RoomDetailAction.EnterQuoteMode) {
-        room.getTimeLineEvent(action.eventId)?.let { timelineEvent ->
-            setState { copy(sendMode = SendMode.QUOTE(timelineEvent, action.text)) }
-        }
-    }
-
-    private fun handleReplyAction(action: RoomDetailAction.EnterReplyMode) {
-        room.getTimeLineEvent(action.eventId)?.let { timelineEvent ->
-            setState { copy(sendMode = SendMode.REPLY(timelineEvent, action.text)) }
-        }
-    }
-
-    private fun handleEnterRegularMode(action: RoomDetailAction.EnterRegularMode) = setState {
-        copy(sendMode = SendMode.REGULAR(action.text, action.fromSharing))
     }
 
     private fun handleOpenOrDownloadFile(action: RoomDetailAction.DownloadOrOpen) {
@@ -1604,7 +1085,7 @@ class RoomDetailViewModel @AssistedInject constructor(
             setState {
                 val typingMessage = typingHelper.getTypingMessage(summary.typingUsers)
                 copy(
-                        typingMessage = typingMessage,
+                        formattedTypingUsers = typingMessage,
                         hasFailedSending = summary.hasFailedSending
                 )
             }
