+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailViewState.kt
@@ -73,7 +73,7 @@ data class RoomDetailViewState(
         val asyncInviter: Async<RoomMemberSummary> = Uninitialized,
         val asyncRoomSummary: Async<RoomSummary> = Uninitialized,
         val activeRoomWidgets: Async<List<Widget>> = Uninitialized,
-        val typingMessage: String? = null,
+        val formattedTypingUsers: String? = null,
         val sendMode: SendMode = SendMode.REGULAR("", false),
         val tombstoneEvent: Event? = null,
         val joinUpgradedRoomAsync: Async<String> = Uninitialized,
@@ -84,7 +84,6 @@ data class RoomDetailViewState(
         val unreadState: UnreadState = UnreadState.Unknown,
         val canShowJumpToReadMarker: Boolean = true,
         val changeMembershipState: ChangeMembershipState = ChangeMembershipState.Unknown,
-        val canSendMessage: Boolean = true,
         val canInvite: Boolean = true,
         val isAllowedToManageWidgets: Boolean = false,
         val isAllowedToStartWebRTCCall: Boolean = true,
