+++ vector/src/main/java/im/vector/app/features/home/room/detail/composer/TextComposerViewState.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2021 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package im.vector.app.features.home.room.detail.composer
+
+import com.airbnb.mvrx.MvRxState
+import im.vector.app.features.home.room.detail.RoomDetailArgs
+import im.vector.app.features.home.room.detail.SendMode
+
+data class TextComposerViewState(
+        val roomId: String,
+        val canSendMessage: Boolean = true,
+        val isSendButtonVisible: Boolean = false,
+        val isRecordingVoiceMessage: Boolean = false,
+        val sendMode: SendMode = SendMode.REGULAR("", false),
+) : MvRxState {
+
+    constructor(args: RoomDetailArgs) : this(
+            roomId = args.roomId,
+    )
+}
+
